late I recore query with test and tidy
