+++ lucene/core/src/java/org/apache/lucene/search/LateInteractionRescorer.java
@@ -6,9 +6,9 @@ import org.apache.lucene.index.VectorSimilarityFunction;
  * Rescores top N results from a first pass query using a {@link LateInteractionFloatValuesSource}
  *
  * <p>Typically, you run a low-cost first pass query to collect results from across the index, then
- * use this rescorer to rerank top N hits using multi-vectors, usually from a late interaction model.
- * Multi-vectors should be indexed in the {@link org.apache.lucene.document.LateInteractionField}
- * provided to rescorer.
+ * use this rescorer to rerank top N hits using multi-vectors, usually from a late interaction
+ * model. Multi-vectors should be indexed in the {@link
+ * org.apache.lucene.document.LateInteractionField} provided to rescorer.
  *
  * @lucene.experimental
  */
@@ -18,9 +18,7 @@ public class LateInteractionRescorer extends DoubleValuesSourceRescorer {
     super(valuesSource);
   }
 
-  /**
-   * Creates a LateInteractionRescorer for provided query vector.
-   */
+  /** Creates a LateInteractionRescorer for provided query vector. */
   public static LateInteractionRescorer create(String fieldName, float[][] queryVector) {
     return create(fieldName, queryVector, VectorSimilarityFunction.COSINE);
   }
@@ -28,16 +26,19 @@ public class LateInteractionRescorer extends DoubleValuesSourceRescorer {
   /**
    * Creates a LateInteractionRescorer for provided query vector.
    *
-   * <p>Top N results from a first pass query are rescored based on the similarity between {@code queryVector} and
-   * the multi-vector indexed in {@code fieldName}. If document does not have a value indexed in {@code fieldName},
-   * a 0f score is assigned.
+   * <p>Top N results from a first pass query are rescored based on the similarity between {@code
+   * queryVector} and the multi-vector indexed in {@code fieldName}. If document does not have a
+   * value indexed in {@code fieldName}, a 0f score is assigned.
    *
-   * @param fieldName the {@link org.apache.lucene.document.LateInteractionField} used for reranking.
+   * @param fieldName the {@link org.apache.lucene.document.LateInteractionField} used for
+   *     reranking.
    * @param queryVector query multi-vector to use for similarity comparison
    * @param vectorSimilarityFunction function used for vector similarity comparisons
    */
-  public static LateInteractionRescorer create(String fieldName, float[][] queryVector, VectorSimilarityFunction vectorSimilarityFunction) {
-    final LateInteractionFloatValuesSource valuesSource = new LateInteractionFloatValuesSource(fieldName, queryVector, vectorSimilarityFunction);
+  public static LateInteractionRescorer create(
+      String fieldName, float[][] queryVector, VectorSimilarityFunction vectorSimilarityFunction) {
+    final LateInteractionFloatValuesSource valuesSource =
+        new LateInteractionFloatValuesSource(fieldName, queryVector, vectorSimilarityFunction);
     return new LateInteractionRescorer(valuesSource);
   }
 
@@ -49,16 +50,19 @@ public class LateInteractionRescorer extends DoubleValuesSourceRescorer {
   /**
    * Creates a LateInteractionRescorer for provided query vector.
    *
-   * <p>Top N results from a first pass query are rescored based on the similarity between {@code queryVector} and
-   * the multi-vector indexed in {@code fieldName}. Falls back to score from the first pass query if a document
-   * does not have a value indexed in {@code fieldName}.
+   * <p>Top N results from a first pass query are rescored based on the similarity between {@code
+   * queryVector} and the multi-vector indexed in {@code fieldName}. Falls back to score from the
+   * first pass query if a document does not have a value indexed in {@code fieldName}.
    *
-   * @param fieldName the {@link org.apache.lucene.document.LateInteractionField} used for reranking.
+   * @param fieldName the {@link org.apache.lucene.document.LateInteractionField} used for
+   *     reranking.
    * @param queryVector query multi-vector to use for similarity comparison
    * @param vectorSimilarityFunction function used for vector similarity comparisons.
    */
-  public static LateInteractionRescorer withFallbackToFirstPassScore(String fieldName, float[][] queryVector, VectorSimilarityFunction vectorSimilarityFunction) {
-    final LateInteractionFloatValuesSource valuesSource = new LateInteractionFloatValuesSource(fieldName, queryVector, vectorSimilarityFunction);
+  public static LateInteractionRescorer withFallbackToFirstPassScore(
+      String fieldName, float[][] queryVector, VectorSimilarityFunction vectorSimilarityFunction) {
+    final LateInteractionFloatValuesSource valuesSource =
+        new LateInteractionFloatValuesSource(fieldName, queryVector, vectorSimilarityFunction);
     return new LateInteractionRescorer(valuesSource) {
       @Override
       protected float combine(float firstPassScore, boolean valuePresent, double sourceValue) {
