+++ lucene/core/src/test/org/apache/lucene/search/TestRescoreTopNQuery.java
@@ -17,18 +17,25 @@
 package org.apache.lucene.search;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.apache.lucene.codecs.lucene99.Lucene99HnswScalarQuantizedVectorsFormat;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.IntField;
 import org.apache.lucene.document.KnnFloatVectorField;
+import org.apache.lucene.document.LateInteractionField;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
+import org.apache.lucene.index.StoredFields;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.store.ByteBuffersDirectory;
@@ -154,6 +161,87 @@ public class TestRescoreTopNQuery extends LuceneTestCase {
         Assert.assertEquals("Score must be 0 for missing DoubleValues", 0, scoreDoc.score, 1e-5);
       }
     }
+  }
+
+  public void testLateInteractionQuery() throws Exception {
+    final String LATE_I_FIELD = "li_vector";
+    final String KNN_FIELD = "knn_vector";
+    List<float[][]> corpus = new ArrayList<>();
+    final int numDocs = atLeast(1000);
+    final int numSegments = random().nextInt(2, 10);
+    final int dim = 128;
+    final VectorSimilarityFunction vectorSimilarityFunction =
+        VectorSimilarityFunction.values()[
+            random().nextInt(VectorSimilarityFunction.values().length)];
+    LateInteractionFloatValuesSource.ScoreFunction scoreFunction =
+        LateInteractionFloatValuesSource.ScoreFunction.values()[
+            random().nextInt(LateInteractionFloatValuesSource.ScoreFunction.values().length)];
+
+    try (Directory dir = newDirectory()) {
+      int id = 0;
+      try (IndexWriter w = new IndexWriter(dir, newIndexWriterConfig())) {
+        for (int j = 0; j < numSegments; j++) {
+          for (int i = 0; i < numDocs; i++) {
+            Document doc = new Document();
+            if (random().nextInt(100) < 30) {
+              // skip value for some docs to create sparse field
+              doc.add(new IntField("has_li_vector", 0, Field.Store.YES));
+            } else {
+              float[][] value = createMultiVector(dim);
+              corpus.add(value);
+              doc.add(new IntField("id", id++, Field.Store.YES));
+              doc.add(new LateInteractionField(LATE_I_FIELD, value));
+              doc.add(new KnnFloatVectorField(KNN_FIELD, randomFloatVector(dim, random())));
+              doc.add(new IntField("has_li_vector", 1, Field.Store.YES));
+            }
+            w.addDocument(doc);
+            w.flush();
+          }
+        }
+        // add a segment with no vectors
+        for (int i = 0; i < 100; i++) {
+          Document doc = new Document();
+          doc.add(new IntField("has_li_vector", 0, Field.Store.YES));
+          w.addDocument(doc);
+        }
+        w.flush();
+      }
+
+      float[][] lateIQueryVector = createMultiVector(dim);
+      float[] knnQueryVector = randomFloatVector(dim, random());
+      KnnFloatVectorQuery knnQuery = new KnnFloatVectorQuery(KNN_FIELD, knnQueryVector, 50);
+
+      try (IndexReader reader = DirectoryReader.open(dir)) {
+        final int topN = 10;
+        IndexSearcher s = new IndexSearcher(reader);
+        TopDocs knnHits = s.search(knnQuery, 5 * topN);
+        Set<Integer> knnHitDocs =
+            Arrays.stream(knnHits.scoreDocs).map(k -> k.doc).collect(Collectors.toSet());
+        Query lateIQuery =
+            RescoreTopNQuery.createLateInteractionQuery(
+                knnQuery, topN, LATE_I_FIELD, lateIQueryVector, vectorSimilarityFunction);
+        TopDocs lateIHits = s.search(lateIQuery, 3 * topN);
+        // total match-set for RescoreTopNQuery is topN
+        assertEquals(topN, lateIHits.scoreDocs.length);
+        StoredFields storedFields = reader.storedFields();
+        for (ScoreDoc hit : lateIHits.scoreDocs) {
+          assertTrue(knnHitDocs.contains(hit.doc));
+          int idValue = Integer.parseInt(storedFields.document(hit.doc).get("id"));
+          float[][] docVector = corpus.get(idValue);
+          float expected =
+              scoreFunction.compare(lateIQueryVector, docVector, vectorSimilarityFunction);
+          assertEquals(expected, hit.score, 1e-5);
+        }
+      }
+    }
+  }
+
+  private float[][] createMultiVector(int dimension) {
+    float[][] value = new float[random().nextInt(3, 12)][];
+    for (int i = 0; i < value.length; i++) {
+      value[i] = randomFloatVector(dimension, random());
+    }
+    return value;
   }
 
   private float[] randomFloatVector(int dimension, Random random) {
