+++ lucene/queries/src/java/org/apache/lucene/queries/function/FunctionScoreQuery.java
@@ -20,14 +20,12 @@ package org.apache.lucene.queries.function;
 import java.io.IOException;
 import java.util.Objects;
 import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.search.DoubleValues;
 import org.apache.lucene.search.DoubleValuesSource;
 import org.apache.lucene.search.Explanation;
 import org.apache.lucene.search.FilterScorer;
 import org.apache.lucene.search.IndexSearcher;
-import org.apache.lucene.search.LateInteractionFloatValuesSource;
 import org.apache.lucene.search.Matches;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.QueryVisitor;
@@ -71,28 +69,6 @@ public final class FunctionScoreQuery extends Query {
    */
   public DoubleValuesSource getSource() {
     return source;
-  }
-
-  /**
-   * Returns a FunctionScoreQuery that re-scores hits from the wrapped query using late-interaction
-   * scores between provided query and indexed document multi-vectors.
-   *
-   * <p>Document multi-vectors are indexed using {@link
-   * org.apache.lucene.document.LateInteractionField}.
-   *
-   * @param in the query to re-score
-   * @param fieldName field containing document multi-vectors for re-scoring
-   * @param queryVector query multi-vector
-   * @param vectorSimilarityFunction vector similarity function used for computing scores
-   */
-  public static FunctionScoreQuery lateInteractionFloatRerankQuery(
-      Query in,
-      String fieldName,
-      float[][] queryVector,
-      VectorSimilarityFunction vectorSimilarityFunction) {
-    LateInteractionFloatValuesSource scoreSource =
-        new LateInteractionFloatValuesSource(fieldName, queryVector, vectorSimilarityFunction);
-    return new FunctionScoreQuery(in, scoreSource);
   }
 
   /**
